requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
freetype=2.9.1=hb4e5f40_0
icu=58.2=h4b95b61_1
jpeg=9b=he5867d9_2
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libiconv=1.15=hdd342a3_7
libpng=1.6.37=ha441bb4_0
libtiff=4.1.0=hcb84e12_0
libxml2=2.9.9=hf6e021a_1
libxslt=1.1.33=h33a18ac_0
lxml=4.5.0=py37hef8c89e_0
ncurses=6.2=h0a44026_0
olefile=0.46=py_0
openssl=1.0.2u=h1de35cc_0
pillow=5.2.0=py37hb68e598_0
pinyin=0.4.0=pypi_0
pip=20.0.2=py37_3
python=3.7.0=hc167b69_0
python-pptx=0.6.18=py_0
python_abi=3.7=1_cp37m
readline=7.0=h1de35cc_5
setuptools=46.2.0=py37_0
sqlite=3.31.1=ha441bb4_0
tk=8.6.7=h35a86e2_3
wheel=0.34.2=py37_0
xlsxwriter=1.2.8=py_0
xz=5.2.4=h1de35cc_4
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
